Install dependencies before the depending.
This appears to be a fix for a regression vs master, as it passes tests.  I
found the issue when working on setup-requires, as in that case a failure
doesn't just break after pip completes, but actually causes the pip invocation
to fail. See issue 2260 in the discussion where its stated that installing
deps first is desired.