Support declarative requirements.
This adds support per the discussion we had on distutils-sig for
declarative requirements in setup.cfg.

Supported are setup-requires (the problem to be solved).
install-requires, needed as a consequences of supporting setup
requirements, because we can't run egg_info during the
pre-installation phase to detect requires. Similarly extras are
supported, for the same reason.

For compatibility with d2to1 we support requires-dist as an alias
for install-requires in the setup.cfg metadata section.